+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming
+### Fixes
+* Fixed spawning way too many Phantoms
+
 ## 4.11.10
 ### Changes
 * Sleeping when tired now skips 1/8 and 1/4 of a day more when Tired II and III